requirements.txt
numpy==1.24.3
pillow==10.4.0
opencv-python==4.8.0.76
opencv-python-headless==4.10.0.84
ultralytics==8.2.63
flask==3.0.3
flask-socketio==5.3.6
python-engineio==4.9.1
Werkzeug==3.0.3
gevent==24.2.1
pandas==2.2.2
wsgiserver==1.3
gevent-websocket==0.10.1
